--- src/app/comment.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { CommentService } from './comment.service';
-
-describe('CommentService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: CommentService = TestBed.get(CommentService);
-    expect(service).toBeTruthy();
-  });
-});
