--- src/app/mock-comments.ts
@@ -1,5 +1,0 @@
-import { Comment } from './comment';
-
-export const COMMENTS: Comment[] = [
-    {maker: 'あんこ', content: '全部同じ問題やん', star: 1},
-];
