+++ src/app/mock/mock-comments.ts
@@ -1,0 +1,5 @@
+import { Comment } from '../models/comment';
+
+export const COMMENTS: Comment[] = [
+    {maker: 'あんこ', content: '全部同じ問題やん', star: 1},
+];
