+++ src/app/models/quiz.ts
