--- src/app/quiz.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { QuizService } from './quiz.service';
-
-describe('QuizService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: QuizService = TestBed.get(QuizService);
-    expect(service).toBeTruthy();
-  });
-});
